test(jest) - First pair of tests written
