+++ src/app/__mock__/articles.js
@@ -1,0 +1,13 @@
+const articles = [
+  {
+    title: 'Hello World',
+    source: 'bbc-news',
+    description: 'This is so cool',
+    url: 'https://newsninja.herokuapp.com',
+    urlToImage: '',
+    author: 'Bamidele Daniel',
+    publishedAt: '',
+  },
+];
+
+export default articles;
