+++ src/app/__mock__/localStorage.js
@@ -1,0 +1,13 @@
+let localStorage = {};
+
+export default {
+  setItem(key, value) {
+    return Object.assign(localStorage, { [key]: value });
+  },
+  getItem(key) {
+    return localStorage[key];
+  },
+  clear() {
+    localStorage = {};
+  }
+};
