--- test/mocha.opts
@@ -1,1 +1,0 @@
--r jsdom-global/register
