+++ tests/mocha.opts
@@ -1,0 +1,1 @@
+-r jsdom-global/register
