+++ tests/stores/store.spec.js
@@ -1,0 +1,14 @@
+// import 'jsdom-global/register';
+// import React from 'react';
+// import { expect } from 'chai';
+// import { shallow, mount } from 'enzyme';
+// import sinon from 'sinon';
+// import NewsStore from '../../src/app/stores/NewsStore';
+
+// describe('Tests for Store', () => {
+//   it('should return an instance of a class', () => {
+//     const spy = sinon.spy(NewsStore);
+//     console.log(spy);
+//   });
+// });
+
