+++ tests/test.js
@@ -1,0 +1,8 @@
+'use strict'
+const expect = require('chai').expect;
+
+describe('Test for True', () => {
+	it('should return true for true', () => {
+		expect(true).to.be.true;
+	});
+});
